Refactor GrpcClient into UnaryGrpcClient and StreamingGrpcClient.
